Fix container's deploy when used with keys.
Fundamentally there was this missing `self.set_common_state(defn)`
that wasn't called on `create` which is essential to retrieve
keys.

This should also fix an issue I had when `storeKeysOnMachine`
was not set explicitly which caused the keys not to be sent.

Attempted to fix the use of `stop`/`start` when there are keys.
There remain some underlying problem I couldn't understand yet.